backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.11
SQLAlchemy==2.0.36
alembic>=1.12.1
pydantic>=2.5.0
email-validator>=2.1.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
jinja2>=3.1.2
pandas>=2.1.3
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
reportlab>=4.0.6
redis>=5.0.1
slowapi>=0.1.9
orjson>=3.9.10
argon2-cffi>=23.1.0
# Production server
gunicorn>=21.2.0